refactored code landing page
