--- src/components/bibleIndex/Book.js
@@ -1,15 +1,0 @@
-import React from 'react'
-
- const Book =({books}) => {
-    return (
-        <div>
-            
-    <select>
-      {Object.keys(books).map(key => (
-        <option value={key}> {books[key]} </option>
-      ))}
-    </select>
-        </div>
-    )
-}
-export default Book;
