--- src/components/bibleIndex/ReadButton.js
@@ -1,10 +1,0 @@
-import React from 'react'
-
- const ReadButton =({label}) => {
-    return (
-        <div>
-           <button>{label}</button> 
-        </div>
-    )
-}
-export default ReadButton;
